seems to be receiving correct commands with timer0 in 16bit mode and Fosc = 1MHz

+++ IRdA.X/tmr_TMR0.h
@@ -19,7 +19,7 @@
 
 
 void tmr_TMR0Init(void);
-uint8_t tmr_computeDelta(uint8_t i);
+uint16_t tmr_computeDelta(uint8_t i);
 uint8_t accquisitionComplete(void);
 void tmr_TMR0mark(void);
 void tmr_TMR0reset(void);
